Completed List view for blueprints
Completed the list view functionality for the blueprints screen. The list view now shows a preview of the image, and is click-able to expand and view. Also added a notification to the blueprints that displays when there are no blueprints to show. This gives better user feedback on what's happening rather than just a white area.