Implemented JWT authentication
Implemented JWT authentication for the GraphQL routes. Now the API will check if the user is presenting an admin token before allowing access to the GrophQL logic.

Not all implementation for this is done.  I still need to implement it on the front end, but the backend logic is in place.